+++ src/graffy/Readme.md
@@ -1,0 +1,3 @@
+# Graffy
+
+Meta package setting up the default Graffy configuration.
