+++ src/graffy/index.js
@@ -1,0 +1,2 @@
+import Graffy from '@graffy/core';
+export default Graffy;
