+++ src/graffy/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "graffy",
+  "version": "1.0.0",
+  "type": "module"
+}
