+++ src/stream/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "@graffy/stream",
+  "version": "1.0.0",
+  "type": "module"
+}
