+++ src/testing/Readme.md
@@ -1,0 +1,3 @@
+# Graffy Testing Utilities
+
+Testing and debugging utilities for Graffy.
